requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
CacheControl==0.12.6
cached-property==1.5.1
certifi==2019.11.28
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-import-export==2.3.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
Django==3.1
django-extensions==3.1.2
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12
drf-yasg==1.20.0
et-xmlfile==1.0.1
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0
httptools==0.1.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
ipaddr==2.2.0
ipython-genutils==0.2.0
ipython==7.16.1
isodate==0.6.0
jdcal==1.4.1
jedi==0.17.2
jsonfield==3.1.0
kavenegar==1.1.2
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests-toolbelt==0.9.1
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
tablib==2.0.0
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8

wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zeep==3.4.0
zipp==3.1.0
zope.interface==5.1.0